remove broken features
