+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
         </activity>
 
         <activity
-            android:name=".ui.photo.PhotoDetailActivity"
+            android:name=".ui.detail.PhotoDetailActivity"
             android:theme="@style/AppTheme.FullScreen" />
 
         <provider
