+++ app/src/main/java/com/semisonfire/cloudgallery/ui/disk/DiskPresenter.kt
@@ -27,7 +27,6 @@ interface DiskPresenter : Presenter {
     fun uploadPhoto(photo: Photo)
     fun getUploadingPhotos()
     fun downloadPhotos(photos: List<Photo>)
-    fun deletePhotos(photos: List<Photo>)
     fun loadMorePhotos()
 
     fun observeContent(): Observable<DiskViewModel>
@@ -53,32 +52,27 @@ class DiskPresenterImpl @Inject constructor(
         compositeDisposable.add(
             uploadManager
                 .uploadListener()
-                .subscribeOn(background())
+                .observeOn(background())
+                .filter { it.uploaded }
                 .subscribe({ uploadResult ->
                     val uploading = viewModel.uploading
                     val uploadedPhoto = uploadResult.photo
-                    if (uploadResult.uploaded) {
-                        if (uploading != null) {
-                            uploading.items as MutableList<UploadItem>
-                            uploading.items.removeAll {
-                                it.id == uploadedPhoto.id
-                            }
-
-                            if (uploading.items.isEmpty()) {
-                                viewModel.uploading = null
-                            }
-
-                            val itemsMap =
-                                mapper.map(listOf(uploadedPhoto), -1)
-
-                            viewModel.mergeItems(itemsMap)
-
-                            diskResultListener.onNext(DiskResult.Update(viewModel.getListItems()))
+                    if (uploading != null) {
+                        uploading.items as MutableList<UploadItem>
+                        uploading.items.removeAll {
+                            it.id == uploadedPhoto.id
                         }
-                    } else {
-                        diskResultListener.onNext(
-                            DiskResult.PhotoUploaded(uploadedPhoto, uploadResult.uploaded)
-                        )
+
+                        if (uploading.items.isEmpty()) {
+                            viewModel.uploading = null
+                        }
+
+                        val itemsMap =
+                            mapper.map(listOf(uploadedPhoto), -1)
+
+                        viewModel.mergeItems(itemsMap)
+
+                        diskResultListener.onNext(DiskResult.Update(viewModel.getListItems()))
                     }
                 }, {
                     it.printThrowable()
@@ -175,19 +169,6 @@ class DiskPresenterImpl @Inject constructor(
                     },
                     { it.printThrowable() }
                 )
-        )
-    }
-
-    override fun deletePhotos(photos: List<Photo>) {
-        compositeDisposable.add(
-            Observable.fromIterable(photos.toMutableList())
-                .concatMap { diskRepository.deletePhoto(it) }
-                .subscribeOn(background())
-                .subscribe({
-                    diskResultListener.onNext(DiskResult.PhotoDeleted(it))
-                }, {
-                    it.printThrowable()
-                })
         )
     }
