+++ app/src/main/java/com/semisonfire/cloudgallery/ui/disk/DiskResult.kt
@@ -2,7 +2,6 @@ package com.semisonfire.cloudgallery.ui.disk
 
 import com.semisonfire.cloudgallery.adapter.holder.Item
 import com.semisonfire.cloudgallery.common.scroll.HorizontalScrollItem
-import com.semisonfire.cloudgallery.data.model.Photo
 import com.semisonfire.cloudgallery.upload.adapter.UploadItem
 
 sealed class DiskResult {
@@ -23,7 +22,5 @@ sealed class DiskResult {
 
     data class Uploading(val uploading: HorizontalScrollItem<UploadItem>) : DiskResult()
 
-    data class PhotoUploaded(val photo: Photo, val uploaded: Boolean) : DiskResult()
     data class PhotoDownloaded(val path: String) : DiskResult()
-    data class PhotoDeleted(val photo: Photo) : DiskResult()
 }
