+++ app/src/main/java/com/semisonfire/cloudgallery/ui/trash/TrashPresenter.kt
@@ -1,12 +1,11 @@
 package com.semisonfire.cloudgallery.ui.trash
 
-import com.semisonfire.cloudgallery.adapter.holder.Item
-import com.semisonfire.cloudgallery.common.photo.PhotoItem
 import com.semisonfire.cloudgallery.core.ui.Presenter
-import com.semisonfire.cloudgallery.data.model.Photo
 import com.semisonfire.cloudgallery.logger.printThrowable
 import com.semisonfire.cloudgallery.ui.disk.LIMIT
+import com.semisonfire.cloudgallery.ui.trash.data.TrashBinMapper
 import com.semisonfire.cloudgallery.ui.trash.data.TrashRepository
+import com.semisonfire.cloudgallery.ui.trash.model.TrashBinResult
 import com.semisonfire.cloudgallery.utils.background
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
@@ -18,19 +17,7 @@ interface TrashPresenter : Presenter {
     fun observeTrashBinResult(): Observable<TrashBinResult>
 
     fun getPhotos(page: Int)
-    fun restorePhotos(photos: List<Photo>)
-    fun deletePhotos(photos: List<Photo>)
     fun clear()
-}
-
-class TrashBinMapper @Inject constructor() {
-
-    fun map(photos: List<Photo>): List<Item> {
-        return photos
-            .asSequence()
-            .map { PhotoItem(it.id, it.name, it.preview) }
-            .toList()
-    }
 }
 
 class TrashPresenterImpl @Inject constructor(
@@ -53,30 +40,6 @@ class TrashPresenterImpl @Inject constructor(
                 .subscribeOn(background())
                 .subscribe(
                     { trashBinResultListener.onNext(TrashBinResult.Loaded(it)) },
-                    { it.printThrowable() }
-                )
-        )
-    }
-
-    override fun restorePhotos(photos: List<Photo>) {
-        compositeDisposable.add(
-            Observable.fromIterable(photos.toMutableList())
-                .concatMap { trashRepository.restoreTrashPhoto(it) }
-                .subscribeOn(background())
-                .subscribe(
-                    { trashBinResultListener.onNext(TrashBinResult.PhotoRestored(it)) },
-                    { it.printThrowable() }
-                )
-        )
-    }
-
-    override fun deletePhotos(photos: List<Photo>) {
-        compositeDisposable.add(
-            Observable.fromIterable(photos.toMutableList())
-                .concatMap { trashRepository.deleteTrashPhoto(it) }
-                .subscribeOn(background())
-                .subscribe(
-                    { trashBinResultListener.onNext(TrashBinResult.PhotoDeleted(it)) },
                     { it.printThrowable() }
                 )
         )
